ai-meetings-full/backend/backend/requirements.txt
fastapi
uvicorn[standard]
openai
#whisper @ git+https://github.com/openai/whisper.git
pydantic
python-multipart
aiofiles
asyncpg
sqlalchemy[asyncio]
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pytz
